fix overflow dection

+++ index.js
@@ -507,7 +507,7 @@ exports.buildModule = function (funcs, imports = [], exports = [], cbs = []) {
   ;; memory related global
   (global $memstart i32  (i32.const 33832))
   ;; the number of 256 words stored in memory
-  (global $wordCount (mut i32) (i32.const 0))
+  (global $wordCount (mut i64) (i64.const 0))
   ;; what was charged for the last memory allocation
   (global $prevMemCost (mut i64) (i64.const 0))
